Move things out of a large index file
